requirements.txt
# GUI Framework
PyQt6>=6.4.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# Deep Learning & Object Detection
ultralytics

# Hardware Communication
pyserial>=3.5

